+++ docs/changes.rst
@@ -16,6 +16,7 @@ Released on ? 2019.
 * Indicate max length and current chars count when editing translation.
 * Improved handling of abbreviations in unchanged translation check.
 * Refreshed landing page for new contributors.
+* Add support for configuring msgmerge addon.
 
 weblate 3.6.1
 -------------